Clean up code a bit.
