+++ aslprep/interfaces/cbf_computation.py
@@ -969,135 +969,3 @@ def refine_ref_mask(t1w_mask, ref_asl_mask, t12ref_transform, tmp_mask, refined_
     results = modify_asl_mask.run()
 
     return results.outputs.out_file
-
-
-class _SplitASLDataInputSpec(BaseInterfaceInputSpec):
-    asl_file = File(exists=True, mandatory=True, desc="ASL file to split.")
-    aslcontext = File(exists=True, mandatory=True, desc="aslcontext TSV.")
-    processing_target = traits.Str()
-    metadata = traits.Dict()
-
-
-class _SplitASLDataOutputSpec(TraitedSpec):
-    asl_file = File(exists=True, desc="Modified ASL file.")
-    aslcontext = File(exists=True, desc="Modified aslcontext file.")
-    metadata = traits.Dict()
-
-
-class SplitASLData(SimpleInterface):
-    """Split ASL data into different files."""
-
-    input_spec = _SplitASLDataInputSpec
-    output_spec = _SplitASLDataOutputSpec
-
-    def _run_interface(self, runtime):
-        aslcontext = pd.read_table(self.inputs.aslcontext)
-        asl_img = nb.load(self.inputs.asl_file)
-        assert asl_img.shape[3] == aslcontext.shape[0]
-
-        if self.inputs.processing_target == "controllabel":
-            files_to_keep = ["control", "label", "m0scan"]
-        elif self.inputs.processing_target == "deltam":
-            files_to_keep = ["deltam", "m0scan"]
-        else:
-            files_to_keep = ["cbf", "m0scan"]
-
-        asl_idx = aslcontext.loc[aslcontext["volume_type"].isin(files_to_keep)].index.values
-        asl_idx = asl_idx.astype(int)
-        self._results["metadata"] = reduce_metadata_lists(self.inputs.metadata, asl_idx)
-
-        asl_img = image.index_img(asl_img, asl_idx)
-
-        self._results["asl_file"] = fname_presuffix(
-            self.inputs.asl_file,
-            suffix="_reduced",
-            newpath=runtime.cwd,
-            use_ext=True,
-        )
-        asl_img.to_filename(self._results["asl_file"])
-
-        aslcontext = aslcontext.loc[asl_idx]
-        self._results["aslcontext"] = fname_presuffix(
-            self.inputs.aslcontext,
-            suffix="_reduced",
-            newpath=runtime.cwd,
-            use_ext=True,
-        )
-        aslcontext.to_csv(self._results["aslcontext"], sep="\t", index=False)
-
-        return runtime
-
-
-def reduce_metadata_lists(metadata, metadata_idx):
-    """Reduce any volume-wise metadata fields to only contain values for selected volumes."""
-    # A hardcoded list of fields that may have one value for each volume.
-    VOLUME_WISE_FIELDS = [
-        "PostLabelingDelay",
-        "VascularCrushingVENC",
-        "LabelingDuration",
-        "EchoTime",
-        "FlipAngle",
-        "RepetitionTimePreparation",
-    ]
-
-    for field in VOLUME_WISE_FIELDS:
-        if field not in metadata:
-            continue
-
-        value = metadata[field]
-        if isinstance(value, list):
-            # Reduce to only the selected volumes
-            metadata[field] = [value[i] for i in metadata_idx]
-
-    return metadata
-
-
-def combine_metadata(metadata, m0scan_metadata, n_asl, n_m0):
-    """Combine any volume-wise metadata fields from the ASL and m0scan dictionaries."""
-    VOLUME_WISE_FIELDS = [
-        "PostLabelingDelay",
-        "VascularCrushingVENC",
-        "LabelingDuration",
-        "EchoTime",
-        "FlipAngle",
-        "RepetitionTimePreparation",
-    ]
-
-    if n_m0 == 0:
-        return metadata
-
-    for field in VOLUME_WISE_FIELDS:
-        if (field not in metadata) and (field not in m0scan_metadata):
-            continue
-        elif (field not in metadata) and not isinstance(m0scan_metadata[field], list):
-            metadata[field] = m0scan_metadata
-            continue
-        elif field not in metadata:
-            raise ValueError(
-                f"{field} is a list in the m0scan metadata, but not defined in the ASL metadata."
-            )
-        elif (field not in m0scan_metadata) and isinstance(metadata[field], list):
-            raise ValueError(
-                f"{field} is a list in the ASL metadata, but not defined in the m0scan metadata."
-            )
-        elif field not in m0scan_metadata:
-            continue
-
-        asl_value = metadata[field]
-        m0scan_value = m0scan_metadata[field]
-        if isinstance(asl_value, list) and isinstance(m0scan_value, list):
-            value = asl_value + m0scan_value
-        elif isinstance(asl_value, list):
-            m0scan_value = [m0scan_value] * n_m0
-            value = asl_value + m0scan_value
-        elif isinstance(m0scan_value, list):
-            asl_value = [asl_value] * n_asl
-            value = asl_value + m0scan_value
-        elif asl_value != m0scan_value:
-            asl_value = [asl_value] * n_asl
-            m0scan_value = [m0scan_value] * n_m0
-            value = asl_value + m0scan_value
-
-        metadata[field] = value
-
-    return metadata
